Move CellReference from hssf.util into ss.util, but leave a class in the old package to proxy requests on. Also needed tweaks to several things that used it. Finally, add lots more tests, and fix the high column number num<->letter conversion
git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@645143 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/model/CommentShape.java
@@ -74,7 +74,7 @@ public class CommentShape extends TextboxShape {
      private NoteRecord createNoteRecord( HSSFComment shape, int shapeId )
     {
         NoteRecord note = new NoteRecord();
-        note.setColumn(shape.getColumn());
+        note.setColumn((short)shape.getColumn());
         note.setRow((short)shape.getRow());
         note.setFlags(shape.isVisible() ? NoteRecord.NOTE_VISIBLE : NoteRecord.NOTE_HIDDEN);
         note.setShapeId((short)shapeId);
